feat(Grid): a grid instance is now an object with 0 or more hexes
Grid's methods are moved to the Grid prototype and the shape methods return the grid instance.

Grid() now returns an actual instance with methods in its prototype. Shape method (e.g.
`Grid#rectangle`) now add hexes to the grid instance and return the grid instance. The grid instance
is an object with serialized hexes (serialization is done using `Hex#toString`) as keys and the
hexes themselves as values.